+++ src/main/java/org/jchess/App.java
@@ -3,6 +3,7 @@ package org.jchess;
 import java.util.Scanner;
 
 import org.jchess.control.BoardManager;
+import org.jchess.exceptions.PieceNotFoundException;
 import org.jchess.model.Board;
 import org.jchess.model.Position;
 import org.jchess.view.UI;
@@ -17,10 +18,16 @@ public class App
 {
     public static void main(String[] args)
     {
-        Board board = BoardManager.generateBoard("r1bqkbnr/pp1p1ppp/2n1p3/2p5/3PP3/2N5/PPP1NPPP/R1BQKB1R b KQkq - 0 4");
+        startGame();
+    }
+
+    public static void startGame()
+    {
+        Board board = BoardManager.generateBoard("8/8/8/8/8/8/8/R3K2R w - - 0 1");
         Scanner scanner = new Scanner(System.in);
 
         UI.printBoard(board);
+        UI.printBoardWithValidMoves(board, BoardManager.getPieceAtPosition(board, new Position("e1")));
 
         while (true)
         {
@@ -29,7 +36,16 @@ public class App
             System.out.print("Your input: ");
             input = scanner.nextLine();
 
-            BoardManager.movePiece(board, input);
+            try
+            {
+                BoardManager.movePiece(board, input);
+            }
+            catch (PieceNotFoundException exception)
+            {
+                System.out.println("There is no piece at the specified position!");
+                continue;
+            }
+
             UI.printBoard(board);
         }
     }