+++ src/main/java/org/jchess/model/Board.java
@@ -8,6 +8,7 @@ public class Board
     // A list of all the pieces on the board
     private Piece[] pieces = new Piece[0];
     private Color playingSideColor = Color.WHITE;
+    private CastlingStatus[] castlingStatuses = { CastlingStatus.KINGANDQUEENSIDE, CastlingStatus.KINGANDQUEENSIDE };
 
     public void setPieces(Piece[] pieces)
     {
@@ -27,5 +28,35 @@ public class Board
     public void setPlayingSideColor(Color playingSideColor)
     {
         this.playingSideColor = playingSideColor;
+    }
+
+    public CastlingStatus[] getCastlingStatuses()
+    {
+        return this.castlingStatuses;
+    }
+
+    public void setCastlingStatuses(CastlingStatus[] castlingStatuses)
+    {
+        this.castlingStatuses = castlingStatuses;
+    }
+
+    public CastlingStatus getWhiteCastlingStatus()
+    {
+        return this.castlingStatuses[0];
+    }
+
+    public void setWhiteCastlingStatus(CastlingStatus castlingStatus)
+    {
+        this.castlingStatuses[0] = castlingStatus;
+    }
+
+    public CastlingStatus getBlackCastlingStatus()
+    {
+        return this.castlingStatuses[1];
+    }
+
+    public void setBlackCastlingStatus(CastlingStatus castlingStatus)
+    {
+        this.castlingStatuses[1] = castlingStatus;
     }
 }