+++ src/main/java/org/jchess/model/CastlingStatus.java
@@ -1,0 +1,9 @@
+package org.jchess.model;
+
+public enum CastlingStatus
+{
+    NONE,
+    KINGSIDE,
+    QUEENSIDE,
+    KINGANDQUEENSIDE
+}